fix: sanitize user input at write time instead of in email templates
Move HTML escaping to a shared sanitize.mjs module and apply it in
create-listing and update-listing before writing to DynamoDB, so data
is safe everywhere it's read (UI, emails, etc.) rather than escaped
per-output-surface.

Also update claim expiry from 24h to 2h with a 30-minute warning window.
